move MetricsObjective to another file (separate from Objetive)
