+++ test/unit/optimizer/test_data_objective_eval.py
@@ -12,7 +12,7 @@ from fedot.core.data.supplementary_data import SupplementaryData
 from fedot.core.optimisers.fitness import SingleObjFitness
 from fedot.core.optimisers.objective import Objective, PipelineObjectiveEvaluate
 from fedot.core.optimisers.objective.data_source_splitter import DataSourceSplitter
-from fedot.core.optimisers.objective.objective import MetricsObjective
+from fedot.core.optimisers.objective.metrics_objective import MetricsObjective
 from fedot.core.pipelines.pipeline import Pipeline
 from fedot.core.pipelines.pipeline_builder import PipelineBuilder
 from fedot.core.repository.dataset_types import DataTypesEnum
