+++ test/unit/optimizer/test_generation_keeper.py
@@ -4,7 +4,8 @@ from fedot.core.optimisers.archive import GenerationKeeper
 from fedot.core.optimisers.fitness import Fitness, MultiObjFitness, null_fitness
 from fedot.core.optimisers.gp_comp.operators.operator import PopulationT
 from fedot.core.optimisers.graph import OptGraph, OptNode
-from fedot.core.optimisers.objective.objective import Objective, MetricsObjective
+from fedot.core.optimisers.objective.objective import Objective
+from fedot.core.optimisers.objective.metrics_objective import MetricsObjective
 from fedot.core.optimisers.opt_history_objects.individual import Individual
 from fedot.core.repository.quality_metrics_repository import ComplexityMetricsEnum, RegressionMetricsEnum
